Move sync hooks to a separate file to avoid circular dependencies and console errors in the tests
